Generate random samples from the inverse Gaussian distribution

This function generates random samples from the inverse Gaussian distribution.

Usage

r_ig(lots, n, mu, lambda)

Arguments

lots

The number of lots to generate.

n

The number of observations per lot.

mu

The mean parameter of the distribution.

lambda

The shape parameter of the distribution.

Value

A matrix of random samples from the inverse Gaussian distribution.

Examples

r_ig(lots = 5, n = 10, mu = 1, lambda = 2)
#>                n_1       n_2       n_3       n_4       n_5       n_6       n_7
#> sample_1 1.0050943 0.7873960 1.5042920 0.4269870 0.2295919 1.4747385 0.8265366
#> sample_2 0.7054206 1.0184362 1.1573637 0.8549261 0.6013737 0.8929825 0.3801120
#> sample_3 0.5739291 1.7375964 0.4238163 0.5510535 0.4172748 0.9157785 0.8032292
#> sample_4 0.7697856 0.5985978 0.9538095 0.6738390 0.2873346 0.4383482 0.7231921
#> sample_5 0.5839989 1.4426244 1.4620014 0.6982063 0.5250957 0.8865185 0.2374699
#>                n_8       n_9      n_10
#> sample_1 0.7028494 0.5927092 0.7823043
#> sample_2 1.0758734 0.8965829 0.3307299
#> sample_3 0.9073973 1.3705492 1.2112966
#> sample_4 0.3992032 2.1606694 1.1228833
#> sample_5 0.7277074 0.9590061 0.3329178