Generate random samples from the inverse Gaussian distribution

This function generates random samples from the inverse Gaussian distribution.

r_ig(lots, n, mu, lambda, lotes = TRUE)

Arguments

lots

The number of lots to generate.

n

The number of observations per lot.

mu

The mean parameter of the distribution.

lambda

The shape parameter of the distribution.

lotes

Logical indicating whether to return the lots as separate matrices.

Value

A matrix of random samples from the inverse Gaussian distribution.

Examples

r_ig(lots = 5, n = 10, mu = 1, lambda = 2)
#>                n_1       n_2       n_3       n_4       n_5       n_6       n_7
#> sample_1 0.8532847 1.2666518 0.8298538 0.5460080 1.6904420 0.7171381 0.9920116
#> sample_2 0.4024114 0.4708130 0.3956621 0.4869686 1.0060061 0.7834631 0.6605191
#> sample_3 1.1541454 0.4327299 1.1035354 0.6154232 0.7582162 0.5730738 0.7390201
#> sample_4 0.7723864 0.7390862 0.3056683 0.7147504 1.5325020 0.9616707 0.3483250
#> sample_5 2.1489346 0.8411871 0.7662106 0.7797205 0.5314632 1.4242220 0.9885492
#>                n_8       n_9      n_10
#> sample_1 1.6992601 0.9577585 0.7677746
#> sample_2 1.2255478 2.0989247 1.2074097
#> sample_3 0.3131799 3.1809018 0.1737116
#> sample_4 0.3348297 0.4974632 1.3779566
#> sample_5 1.4357818 0.5665660 0.5124337