Calculate statistical measures for a Beta Prime control chart

Calculate statistical measures for a Beta Prime control chart

stats_bp(data, alpha = 0.0027, mu = NULL, phi = NULL, ...)

Arguments

data

A matrix of observed data

alpha

The significance level for the control limits

mu

The mean parameter of the Beta Prime distribution (optional)

phi

The shape parameter of the Beta Prime distribution (optional)

...

Additional arguments to be passed to limits_bp

Value

A list containing the estimated parameters, control limits, and statistical measures

Examples

set.seed(0)
x <- r_bp(lots = 1000, n = 10, mu = 1, phi = 2)

# Estimate for maximum likelihood
stats_bp(data = x, alpha = 0.0027)
#> $mu_hat
#> [1] 1.007516
#> 
#> $phi_hat
#> [1] 1.930415
#> 
#> $alpha_hat
#> [1] 0.007
#> 
#> $ARL
#> [1] 142.8571
#> 
#> $MRL
#> [1] 98.67405
#> 
#> $SDRL
#> [1] 142.3563
#> 
#> $li
#> [1] 0.4248392
#> 
#> $ls
#> [1] 2.646655
#> 
#> $convergence
#> [1] -994
#> 
#> $value
#> [1] 175.5181
#> 

# Useing the true parameters
stats_bp(data = x, alpha = 0.0027, mu = 1, phi = 2)
#> $mu_hat
#> [1] 1
#> 
#> $phi_hat
#> [1] 2
#> 
#> $alpha_hat
#> [1] 0.007
#> 
#> $ARL
#> [1] 142.8571
#> 
#> $MRL
#> [1] 98.67405
#> 
#> $SDRL
#> [1] 142.3563
#> 
#> $li
#> [1] 0.4265174
#> 
#> $ls
#> [1] 2.58941
#> 
#> $convergence
#> NULL
#> 
#> $value
#> NULL
#>